refactor: extract fetching logic
